refactor(multi-select): cleanup types
